[Feat/#28] 연결뷰-뷰와 뷰모델 파일 분리
